pyproject.toml
[tool.poetry]
name = "cloud-worker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
google-cloud-storage = "^2.8.0"
google-cloud-pubsub = "^2.17.0"
flask = "2.2.3"
gunicorn = "20.1.0"
cloud-db = {git = "https://github.com/b-munar/cloud-db.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
